Preserve the domid when swapping 16 bit grant table entry flags
We use an atomic CMPXCHG on first 32 bits of the grant table entry
when revoking access to the memory page.  Target domain ID field is
part of these 32 bits, thus shouldn't be masked out for comparison.
This appears to be the last piece of the QubesOS VM chaining puzzle;
tested by Marco Peereboom, thanks!